directory refactoring + mapping + mapstruct spring settings
