+++ customer-service/src/main/java/com/learning/microservices/customer/api/dto/CustomerRequest.java
@@ -1,6 +1,6 @@
 package com.learning.microservices.customer.api.dto;
 
-public record CustomerRegistatrationDto(
+public record CustomerRequest(
         String firstName,
         String lastName,
         String email
