+++ customer-service/src/main/java/com/learning/microservices/customer/api/dto/CustomerWebDto.java
@@ -1,33 +1,13 @@
 package com.learning.microservices.customer.api.dto;
 
 import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
-@Getter
-@Setter
-@Builder
+@Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class CustomerWebDto {
     Integer id;
     String email;
     String fullName;
-    // todo add String fullName (concat firstName + lastName) via MapStruct - СДЕЛАНО
-
-
-    public CustomerWebDto(Integer id, String email, String fullName) {
-        this.id = id;
-        this.email = email;
-        this.fullName = fullName;
-    }
-
-    public CustomerWebDto(Integer id, String email) {
-        this.id = id;
-        this.email = email;
-    }
-
-    public CustomerWebDto() {
-    }
 }
