+++ customer-service/src/main/java/com/learning/microservices/customer/api/mapper/CustomerWebMapper.java
@@ -1,25 +1,16 @@
 package com.learning.microservices.customer.api.mapper;
 
 import com.learning.microservices.customer.api.dto.CustomerWebDto;
-import com.learning.microservices.customer.store.entity.Customer;
-import org.mapstruct.AfterMapping;
+import com.learning.microservices.customer.domain.entity.Customer;
 import org.mapstruct.Mapper;
-import org.mapstruct.MappingTarget;
+import org.mapstruct.Mapping;
 
 import java.util.List;
 
-@Mapper(componentModel = "spring")
+@Mapper
 public interface CustomerWebMapper {
+    @Mapping(target = "fullName", source = "fullName")
     CustomerWebDto dto(Customer customer);
 
     List<CustomerWebDto> dtos(List<Customer> customers);
-
-    @AfterMapping
-    default void setFullName(Customer customer,
-                             @MappingTarget CustomerWebDto customerWebDto) {
-        customerWebDto.setFullName(customer.getFirstName()
-                + " "
-                + customer.getLastName()
-        );
-    }
 }
