+++ customer-service/src/main/java/com/learning/microservices/customer/domain/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
-package com.learning.microservices.customer.store.repository;
+package com.learning.microservices.customer.domain.repository;
 
-import com.learning.microservices.customer.store.entity.Customer;
+import com.learning.microservices.customer.domain.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
