+++ customer-service/src/main/java/com/learning/microservices/customer/service/impl/CustomerServiceImpl.java
@@ -4,11 +4,11 @@ import com.learning.microservices.amqp.RabbitMQMessageProducer;
 import com.learning.microservices.clients.fraud.FraudServiceClient;
 import com.learning.microservices.clients.fraud.dto.FraudCheckResponse;
 import com.learning.microservices.clients.notification.dto.NotificationRequest;
-import com.learning.microservices.customer.api.dto.CustomerRegistatrationDto;
+import com.learning.microservices.customer.api.dto.CustomerRequest;
+import com.learning.microservices.customer.domain.entity.Customer;
+import com.learning.microservices.customer.domain.repository.CustomerRepository;
 import com.learning.microservices.customer.service.CustomerService;
-import com.learning.microservices.customer.service.mapper.CustomerServiceMapper;
-import com.learning.microservices.customer.store.entity.Customer;
-import com.learning.microservices.customer.store.repository.CustomerRepository;
+import com.learning.microservices.customer.service.mapper.CustomerMapper;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
@@ -27,17 +27,15 @@ public class CustomerServiceImpl implements CustomerService {
     FraudServiceClient fraudServiceClient;
     //    NotificationClient notificationClient;
     RabbitMQMessageProducer rabbitMessageProducer;
-    CustomerServiceMapper customerServiceMapper;
-    // todo create and inject new CustomerMapper - СДЕЛАНО
+    CustomerMapper mapper;
 
     @Override
-    public void save(CustomerRegistatrationDto request) {
-        // todo create empty Customer - СДЕЛАНО
-        // todo update empty Customer via new CustomerMapper (dont update id) - СДЕЛАНО
-        Customer customer = customerServiceMapper.toEntity(request);
+    public void save(CustomerRequest request) {
+        Customer customer = new Customer();
+        mapper.update(customer, request);
         // todo email validation
         // todo email not taken
-        
+
         repository.saveAndFlush(customer);
         if (isFraudster(customer)) {
             throw new IllegalStateException("fraudster");
