+++ pom.xml
@@ -165,6 +165,7 @@
                             <groupId>org.mapstruct</groupId>
                             <artifactId>mapstruct-processor</artifactId>
                             <version>${mapstruct.version}</version>
+                            <scope>provided</scope>
                         </path>
                         <path>
                             <groupId>org.projectlombok</groupId>
@@ -172,6 +173,9 @@
                             <version>${lombok.version}</version>
                         </path>
                     </annotationProcessorPaths>
+                    <compilerArgs>
+                        <arg>-Amapstruct.defaultComponentModel=spring</arg>
+                    </compilerArgs>
                     <source>21</source>
                     <target>21</target>
                 </configuration>
